Add nonstd::span
Added span to allow removal of output parameters (but still without
allocation) and cleaner passing of (sub)spans.
Note: To be honest really want an abstraction that is a
std::array+std::span where you can pop back but just decrease the span
ptr. Although I note that cppreference docs does not seem to have a way
to take a current span and change the size, only build a new one ie.
the spans are immutable which has its benefits I guess.